+++ vscode-extension/extension.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const assert = require('assert');
-const { route } = require('./extension');
+const { route, smartPaste } = require('./extension');
 
 (async () => {
   const calls = [];
@@ -15,6 +15,21 @@ const { route } = require('./extension');
 
   vscode.window.activeTerminal = undefined;
   assert.doesNotThrow(() => route(vscode, '\u001b[1;5H'));
+
+  const commandCalls = [];
+  vscode.commands = {
+    executeCommand: async (...args) => commandCalls.push(args)
+  };
+  vscode.env = { clipboard: { readText: async () => 'clipboard text' } };
+  await smartPaste(vscode);
+  assert.deepStrictEqual(commandCalls.pop(), ['workbench.action.terminal.paste']);
+
+  vscode.env.clipboard.readText = async () => '';
+  await smartPaste(vscode);
+  assert.deepStrictEqual(commandCalls.pop(), [
+    'workbench.action.terminal.sendSequence',
+    { text: '\u0016' }
+  ]);
   console.log('VS Code bridge tests passed');
 })().catch((error) => {
   console.error(error);